image-repair-backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
numpy==1.24.3
Pillow==10.1.0
rembg==2.0.50
gunicorn==21.2.0
Werkzeug==3.0.1
